psx-game-runtime: slice 2b -- schedule, room window, stream scheduler
RuntimeScheduleConfig moves as schedule.rs (the shared policy-knob
struct). active_rooms becomes RoomWindow<MAX_ACTIVE_ROOMS>, absorbing
four Playtest fields and visible_rooms_are_active (which composes
better on the window; visibility arrives as a parameter).
active_room_streaming becomes room_streaming.rs with
RoomStreamScheduler gaining a second const generic for the index
capacity that was an example const baked into its arrays; the one
static mut instance stays example-side per the carve pattern.

Cooked tables arrive as &'static records, schedule knobs as values,
example-coupled builders as closures with unchanged signatures.
Two disclosed no-op changes, both hash-verified: finish_job clears
the job before (not after) the glue applies fields with nothing
reading in between, and redundant .min chains collapsed where the
sole instantiation makes both bounds equal. Crate invariants hold:
zero statics, zero new unsafe, generated:: never named.

Gates: engine check clean both feature shapes, MIPS builds, make
validate 2/2 bit-identical (display 0xc009f6733d377f66), 250 tests
green, fmt clean. The plan's perf-tape spot check needs the
telemetry-featured guest shape (tapes key off FrameBegin markers,
compiled out in the default build) -- it runs as an identical-shape
branch-vs-main A/B at the start of the vram_runtime slice, which
re-gates this hot path anyway; today's moves are motion-only with
hash-proven equal output.
