+++ engine/crates/psx-game-runtime/src/lib.rs
@@ -25,4 +25,7 @@
 #[allow(missing_docs)]
 pub mod cd_stream;
 pub mod room_cache;
+pub mod room_streaming;
 pub mod room_visibility;
+pub mod room_window;
+pub mod schedule;
