+++ engine/crates/psx-game-runtime/src/room_window.rs
@@ -1,0 +1,301 @@
+//! Active-room window state machine, carved out of `editor-playtest`'s
+//! `active_rooms` module (phase 1, slice 2 of
+//! docs/game-runtime-plan.md). [`RoomWindow`] owns the resident draw
+//! window, the incremental rebuild job staged against it, and the
+//! request anchor the example previously spread across its scene
+//! struct; cooked tables arrive as `&'static` psx-level records,
+//! capacities as `const N` generic parameters, and the
+//! streaming/residency-coupled room builders stay with the game as
+//! closures until those modules move.
+
+use crate::room_cache::{
+    active_room_contains_drawable, ActiveRoomWindowJob, ActiveRuntimeRoom, INVALID_ROOM_INDEX,
+};
+use crate::room_visibility::RoomVisibility;
+use psx_engine::{telemetry, RoomPoint};
+use psx_level::{LevelRoomRecord, RoomIndex};
+
+/// Outcome of one [`RoomWindow::step_job`] build slice.
+pub struct RoomWindowStep {
+    /// Rooms built (or reused) into the staged window this slice.
+    pub built: usize,
+    /// Requested room that failed to build, or [`INVALID_ROOM_INDEX`].
+    pub unbuilt_room: RoomIndex,
+    /// The freshly built current room, when this slice produced it.
+    pub current_active: Option<ActiveRuntimeRoom>,
+}
+
+/// Owned active-room window state: the resident draw window, the
+/// incremental rebuild job staged against it, the player anchor the
+/// latest window request ran for, and the skipped-build diagnostics.
+///
+/// The game supplies its generated budget consts as the generic
+/// parameter and keeps one instance wherever it keeps scene state.
+pub struct RoomWindow<const MAX_ACTIVE_ROOMS: usize> {
+    /// Cache-budgeted draw chunks, all expressed relative to the game's
+    /// current room.
+    pub rooms: [Option<ActiveRuntimeRoom>; MAX_ACTIVE_ROOMS],
+    /// Incremental active-room cache rebuild in progress. The old
+    /// `rooms` remain drawable until the staged replacement is ready.
+    pub job: ActiveRoomWindowJob<MAX_ACTIVE_ROOMS>,
+    /// Player position the latest window request was anchored to.
+    pub anchor: RoomPoint,
+    /// Rooms whose staged build was skipped as not yet drawable.
+    pub cache_skips: u16,
+}
+
+impl<const MAX_ACTIVE_ROOMS: usize> RoomWindow<MAX_ACTIVE_ROOMS> {
+    /// Zeroed state; `const` so the game can keep it in link-time
+    /// zero-initialized storage.
+    pub const EMPTY: Self = Self {
+        rooms: [const { None }; MAX_ACTIVE_ROOMS],
+        job: ActiveRoomWindowJob::EMPTY,
+        anchor: RoomPoint::ZERO,
+        cache_skips: 0,
+    };
+
+    /// Whether the window holds `index` in any slot.
+    pub fn contains(&self, index: RoomIndex) -> bool {
+        let mut slot = 0usize;
+        while slot < MAX_ACTIVE_ROOMS {
+            if self.rooms[slot].is_some_and(|active| active.index == index) {
+                return true;
+            }
+            slot += 1;
+        }
+        false
+    }
+
+    /// Re-express every windowed room's chunk offsets relative to
+    /// `current_record`, dropping rooms whose record or resident stream
+    /// slot (per `stream_slot_for`) is gone.
+    pub fn rebase_to_current_room(
+        &mut self,
+        rooms: &'static [LevelRoomRecord],
+        current_record: &LevelRoomRecord,
+        stream_slot_for: impl Fn(RoomIndex) -> u16,
+    ) {
+        let mut slot = 0usize;
+        while slot < MAX_ACTIVE_ROOMS {
+            let Some(active) = self.rooms[slot] else {
+                slot += 1;
+                continue;
+            };
+            let Some(record) = rooms.get(active.index.to_usize()) else {
+                self.rooms[slot] = None;
+                slot += 1;
+                continue;
+            };
+            if active.stream_slot != stream_slot_for(active.index) {
+                self.rooms[slot] = None;
+                slot += 1;
+                continue;
+            }
+            self.rooms[slot] = Some(active.with_current_room_offsets(record, current_record));
+            slot += 1;
+        }
+    }
+
+    /// Carry still-valid rooms from `previous_rooms` into the free
+    /// window slots after a rebuild, up to `retained_inactive_rooms`
+    /// beyond `*next_slot` (capped by `active_limit`).
+    pub fn retain_previous_rooms(
+        &mut self,
+        rooms: &'static [LevelRoomRecord],
+        previous_rooms: &[Option<ActiveRuntimeRoom>; MAX_ACTIVE_ROOMS],
+        current_record: &LevelRoomRecord,
+        retained_inactive_rooms: usize,
+        active_limit: usize,
+        next_slot: &mut usize,
+        stream_slot_for: impl Fn(RoomIndex) -> u16,
+    ) {
+        let retained_limit = next_slot
+            .saturating_add(retained_inactive_rooms)
+            .min(active_limit)
+            .min(MAX_ACTIVE_ROOMS);
+        let mut previous_slot = 0usize;
+        while *next_slot < retained_limit && previous_slot < MAX_ACTIVE_ROOMS {
+            let Some(previous) = previous_rooms[previous_slot] else {
+                previous_slot += 1;
+                continue;
+            };
+            previous_slot += 1;
+            if previous.stream_slot != stream_slot_for(previous.index)
+                || self.contains(previous.index)
+            {
+                continue;
+            }
+            let Some(record) = rooms.get(previous.index.to_usize()) else {
+                continue;
+            };
+            self.rooms[*next_slot] =
+                Some(previous.with_current_room_offsets(record, current_record));
+            *next_slot += 1;
+        }
+    }
+
+    /// Stage an incremental window rebuild for `current_room` over the
+    /// caller-assembled request list, anchored at `anchor`.
+    pub fn begin_job(
+        &mut self,
+        current_room: RoomIndex,
+        requested_rooms: [RoomIndex; MAX_ACTIVE_ROOMS],
+        requested_count: usize,
+        update_streaming: bool,
+        anchor: RoomPoint,
+    ) {
+        self.anchor = anchor;
+        self.cache_skips = 0;
+        self.job = ActiveRoomWindowJob {
+            active: true,
+            update_streaming,
+            current_room,
+            requested_rooms,
+            requested_count,
+            cursor: 0,
+            next_slot: 0,
+            rooms: [const { None }; MAX_ACTIVE_ROOMS],
+            previous_rooms: self.rooms,
+        };
+        telemetry::counter(telemetry::counter::ROOM_WINDOW_REBUILDS, 1);
+    }
+
+    /// Run up to `active_job_builds_per_tick` staged builds of the
+    /// pending job. `build` is the game's (streaming/residency-coupled)
+    /// room builder over `(slot, index, record, previous_rooms)`;
+    /// `build_blocked` reports whether a failed room is still loading
+    /// (the slice then stops instead of skipping it).
+    pub fn step_job(
+        &mut self,
+        rooms: &'static [LevelRoomRecord],
+        active_job_builds_per_tick: usize,
+        mut build: impl FnMut(
+            usize,
+            RoomIndex,
+            &'static LevelRoomRecord,
+            &[Option<ActiveRuntimeRoom>; MAX_ACTIVE_ROOMS],
+        ) -> Option<ActiveRuntimeRoom>,
+        build_blocked: impl Fn(RoomIndex) -> bool,
+    ) -> RoomWindowStep {
+        let current_room = self.job.current_room;
+        let mut built_this_tick = 0usize;
+        let mut skipped = 0u16;
+        let mut unbuilt_room = INVALID_ROOM_INDEX;
+        let mut current_active = None;
+        {
+            let job = &mut self.job;
+            while job.cursor < job.requested_count
+                && job.next_slot < MAX_ACTIVE_ROOMS
+                && built_this_tick < active_job_builds_per_tick
+            {
+                let index = job.requested_rooms[job.cursor];
+                if index == INVALID_ROOM_INDEX {
+                    job.cursor += 1;
+                    continue;
+                }
+                let Some(record) = rooms.get(index.to_usize()) else {
+                    job.cursor += 1;
+                    continue;
+                };
+                match build(job.next_slot, index, record, &job.previous_rooms) {
+                    Some(active)
+                        if job.cursor == 0
+                            || active.render_room.is_some()
+                            || active.surface_cache.ready =>
+                    {
+                        job.rooms[job.next_slot] = Some(active);
+                        if active.index == current_room {
+                            current_active = Some(active);
+                        }
+                        job.next_slot += 1;
+                        job.cursor += 1;
+                        built_this_tick += 1;
+                    }
+                    Some(_) => {
+                        skipped = skipped.saturating_add(1);
+                        job.cursor += 1;
+                    }
+                    None => {
+                        unbuilt_room = index;
+                        if build_blocked(index) {
+                            break;
+                        }
+                        job.cursor += 1;
+                    }
+                }
+            }
+        }
+        self.cache_skips = self.cache_skips.saturating_add(skipped);
+        RoomWindowStep {
+            built: built_this_tick,
+            unbuilt_room,
+            current_active,
+        }
+    }
+
+    /// When the pending job has consumed its request list (or filled the
+    /// window), promote the staged rooms to the live window, retain
+    /// still-valid previous rooms, and clear the job. Returns whether
+    /// the job landed, so the caller can refresh its current-room state.
+    pub fn finish_job(
+        &mut self,
+        rooms: &'static [LevelRoomRecord],
+        current_record: &LevelRoomRecord,
+        retained_inactive_rooms: usize,
+        active_limit: usize,
+        stream_slot_for: impl Fn(RoomIndex) -> u16,
+    ) -> bool {
+        if self.job.cursor < self.job.requested_count && self.job.next_slot < MAX_ACTIVE_ROOMS {
+            return false;
+        }
+        self.rooms = self.job.rooms;
+        let previous_rooms = self.job.previous_rooms;
+        let mut next_slot = self.job.next_slot;
+        self.retain_previous_rooms(
+            rooms,
+            &previous_rooms,
+            current_record,
+            retained_inactive_rooms,
+            active_limit,
+            &mut next_slot,
+            stream_slot_for,
+        );
+        self.job = ActiveRoomWindowJob::EMPTY;
+        true
+    }
+
+    /// Whether the current room and every frustum-visible room (up to
+    /// `active_chunk_limit`) is present and drawable in the active window.
+    pub fn visible_rooms_are_active<
+        const MAX_PORTAL_FRUSTUMS: usize,
+        const MAX_PORTAL_FRONTIER_ROOMS: usize,
+        const MAX_PORTAL_ROOM_BOUNDS: usize,
+    >(
+        &self,
+        visibility: &RoomVisibility<
+            MAX_ACTIVE_ROOMS,
+            MAX_PORTAL_FRUSTUMS,
+            MAX_PORTAL_FRONTIER_ROOMS,
+            MAX_PORTAL_ROOM_BOUNDS,
+        >,
+        current_room: RoomIndex,
+        active_chunk_limit: usize,
+    ) -> bool {
+        if !active_room_contains_drawable(&self.rooms, current_room, current_room) {
+            return false;
+        }
+        let visible_limit = visibility.visible_room_limit(active_chunk_limit);
+        let mut i = 0usize;
+        while i < visible_limit {
+            if !active_room_contains_drawable(
+                &self.rooms,
+                current_room,
+                visibility.result.rooms[i].room,
+            ) {
+                return false;
+            }
+            i += 1;
+        }
+        true
+    }
+}
